Do not truncate CallstackFrameId hash.
Frame address typically will be 64 bit, but size_t is 32-bit on emscripten.

+++ server/TracyWorker.hpp
@@ -231,7 +231,7 @@ private:
 
     struct CallstackFrameIdHash
     {
-        size_t operator()( const CallstackFrameId& id ) const { return id.data; }
+        uint64_t operator()( const CallstackFrameId& id ) const { return id.data; }
     };
 
     struct CallstackFrameIdCompare
